install html-webpack-plugin
This plugin will generate index.html file in the same directory where our bundle.js is created by Webpack. In this case, the build directory.

Why do we need this plugin? Because Netlify requires a single directory to be made the root directory, so we can’t use index.html in our root directory using Netlify